Adding animations to page transitions between metadata display pages (#47)
* moved changes from prev branch

* removed extra code from scss file

* refactored to usePrevious method for states instead of tracking previous state separately

* removed default imports

* added javadocs and moved type declaration to corrct file

* Revert "removed extra code from scss file"

This reverts commit 582390e32061e2fd0f310cf3d261138cb2c53207.
Reverting in order to debug build errors

* reverted to last working commit to debug build issues

* added in bug fix for react-use library and removed extra scss code

* re-add changes back in after debugging system issue

* fixed buttons, spacing, and prevState formatting

* fixed comment spacing
