+++ src/renderer/index.scss
@@ -36,3 +36,69 @@ td:empty::after,
 .place-held:empty::after {
   content: '\00a0';
 }
+// Global styles
+div#app {
+  width: 100vw;
+  height: 100vh;
+}
+
+.slide-down-appear {
+  opacity: 0;
+  transform: translateY(-100%);
+}
+.slide-down-appear-active {
+  opacity: 1;
+  transform: translateY(0%);
+  transition: opacity 300ms, transform 300ms;
+}
+
+.slide-left-appear {
+  opacity: 0;
+  transform: translateX(100%);
+}
+.slide-left-appear-active {
+  opacity: 1;
+  transform: translateX(0%);
+  transition: opacity 300ms, transform 300ms;
+}
+
+.slide-right-appear {
+  opacity: 0;
+  transform: translateX(-100%);
+}
+.slide-right-appear-active {
+  opacity: 1;
+  transform: translateX(0%);
+  transition: opacity 300ms, transform 300ms;
+}
+.config-rec-set-appear {
+  opacity: 0;
+  transform: translateY(100%);
+}
+.config-rec-set-appear-active {
+  opacity: 1;
+  transform: translateY(0%);
+  transition: opacity 300ms, transform 300ms;
+}
+.config-rec-set-exit {
+  opacity: 1;
+  transform: translateY(0%);
+}
+.config-rec-set-exit-active {
+  opacity: 0;
+  transform: translateY(-100%);
+  transition: opacity 300ms, transform 300ms;
+}
+
+// Unless specified otherwise, all images are not draggable and not selectable
+* {
+  -webkit-user-drag: none;
+  user-select: none;
+}
+
+// https://stackoverflow.com/a/42225797
+td:empty::after,
+.list-group-item:empty::after,
+.place-held:empty::after {
+  content: '\00a0';
+}
