+++ src/renderer/types/global.d.ts
@@ -1,1 +1,3 @@
 declare const __static: string;
+// Bug fix for ts error in react-use
+type PositionError = GeolocationPositionError;
